chore(hummock-manager): reorganize hummock manager methods
