+++ src/meta/src/hummock/manager/versioning.rs
@@ -16,21 +16,24 @@ use std::cmp;
 use std::collections::{BTreeMap, HashMap, HashSet};
 
 use itertools::Itertools;
+use risingwave_common::catalog::TableId;
 use risingwave_common::util::epoch::INVALID_EPOCH;
 use risingwave_hummock_sdk::compaction_group::hummock_version_ext::{
-    build_initial_compaction_group_levels, get_compaction_group_ids, BranchedSstInfo,
+    build_initial_compaction_group_levels, get_compaction_group_ids,
+    get_table_compaction_group_id_mapping, BranchedSstInfo,
 };
-use risingwave_hummock_sdk::compaction_group::StaticCompactionGroupId;
+use risingwave_hummock_sdk::compaction_group::{StateTableId, StaticCompactionGroupId};
 use risingwave_hummock_sdk::table_stats::add_prost_table_stats_map;
 use risingwave_hummock_sdk::version::{HummockVersion, HummockVersionDelta};
 use risingwave_hummock_sdk::{
-    CompactionGroupId, HummockContextId, HummockSstableObjectId, HummockVersionId, FIRST_VERSION_ID,
+    CompactionGroupId, HummockContextId, HummockEpoch, HummockSstableObjectId, HummockVersionId,
+    FIRST_VERSION_ID,
 };
 use risingwave_pb::common::WorkerNode;
 use risingwave_pb::hummock::write_limits::WriteLimit;
 use risingwave_pb::hummock::{
-    CompactionConfig, HummockPinnedSnapshot, HummockPinnedVersion, HummockVersionStats,
-    SstableInfo, TableStats,
+    CompactionConfig, HummockPinnedSnapshot, HummockPinnedVersion, HummockSnapshot,
+    HummockVersionStats, SstableInfo, TableStats,
 };
 use risingwave_pb::meta::subscribe_response::{Info, Operation};
 
@@ -40,34 +43,11 @@ use crate::hummock::manager::checkpoint::HummockVersionCheckpoint;
 use crate::hummock::manager::commit_multi_var;
 use crate::hummock::manager::context::ContextInfo;
 use crate::hummock::manager::gc::DeleteObjectTracker;
-use crate::hummock::manager::worker::{HummockManagerEvent, HummockManagerEventSender};
-use crate::hummock::metrics_utils::{
-    trigger_safepoint_stat, trigger_write_stop_stats, LocalTableMetrics,
-};
+use crate::hummock::metrics_utils::{trigger_write_stop_stats, LocalTableMetrics};
 use crate::hummock::model::CompactionGroup;
 use crate::hummock::HummockManager;
 use crate::model::VarTransaction;
-use crate::storage::MetaStore;
 use crate::MetaResult;
-
-/// `HummockVersionSafePoint` prevents hummock versions GE than it from being GC.
-/// It's used by meta node itself to temporarily pin versions.
-pub struct HummockVersionSafePoint {
-    pub id: HummockVersionId,
-    event_sender: HummockManagerEventSender,
-}
-
-impl Drop for HummockVersionSafePoint {
-    fn drop(&mut self) {
-        if self
-            .event_sender
-            .send(HummockManagerEvent::DropSafePoint(self.id))
-            .is_err()
-        {
-            tracing::debug!("failed to drop hummock version safe point {}", self.id);
-        }
-    }
-}
 
 #[derive(Default)]
 pub struct Versioning {
@@ -157,29 +137,51 @@ impl HummockManager {
         Ok(workers)
     }
 
+    /// Gets current version without pinning it.
+    /// Should not be called inside [`HummockManager`], because it requests locks internally.
+    ///
+    /// Note: this method can hurt performance because it will clone a large object.
+    pub async fn get_current_version(&self) -> HummockVersion {
+        self.versioning.read().await.current_version.clone()
+    }
+
+    pub async fn get_current_max_committed_epoch(&self) -> HummockEpoch {
+        self.versioning
+            .read()
+            .await
+            .current_version
+            .max_committed_epoch
+    }
+
+    /// Gets the mapping from table id to compaction group id
+    pub async fn get_table_compaction_group_id_mapping(
+        &self,
+    ) -> HashMap<StateTableId, CompactionGroupId> {
+        get_table_compaction_group_id_mapping(&self.versioning.read().await.current_version)
+    }
+
+    /// Get version deltas from meta store
+    #[cfg_attr(coverage, coverage(off))]
+    pub async fn list_version_deltas(
+        &self,
+        start_id: u64,
+        num_limit: u32,
+        committed_epoch_limit: HummockEpoch,
+    ) -> Result<Vec<HummockVersionDelta>> {
+        let versioning = self.versioning.read().await;
+        let version_deltas = versioning
+            .hummock_version_deltas
+            .range(start_id..)
+            .map(|(_id, delta)| delta)
+            .filter(|delta| delta.max_committed_epoch <= committed_epoch_limit)
+            .take(num_limit as _)
+            .cloned()
+            .collect();
+        Ok(version_deltas)
+    }
+
     pub async fn get_version_stats(&self) -> HummockVersionStats {
         self.versioning.read().await.version_stats.clone()
-    }
-
-    pub async fn register_safe_point(&self) -> HummockVersionSafePoint {
-        let versioning = self.versioning.read().await;
-        let mut wl = self.context_info.write().await;
-        let safe_point = HummockVersionSafePoint {
-            id: versioning.current_version.id,
-            event_sender: self.event_sender.clone(),
-        };
-        wl.version_safe_points.push(safe_point.id);
-        trigger_safepoint_stat(&self.metrics, &wl.version_safe_points);
-        safe_point
-    }
-
-    pub async fn unregister_safe_point(&self, safe_point: HummockVersionId) {
-        let mut wl = self.context_info.write().await;
-        let version_safe_points = &mut wl.version_safe_points;
-        if let Some(pos) = version_safe_points.iter().position(|sp| *sp == safe_point) {
-            version_safe_points.remove(pos);
-        }
-        trigger_safepoint_stat(&self.metrics, &wl.version_safe_points);
     }
 
     /// Updates write limits for `target_groups` and sends notification.
@@ -244,6 +246,46 @@ impl HummockManager {
         version_stats.table_stats = new_stats.table_stats;
         commit_multi_var!(self.meta_store_ref(), version_stats)?;
         Ok(())
+    }
+
+    pub fn latest_snapshot(&self) -> HummockSnapshot {
+        let snapshot = self.latest_snapshot.load();
+        HummockSnapshot::clone(&snapshot)
+    }
+
+    /// We don't commit an epoch without checkpoint. We will only update the `max_current_epoch`.
+    pub fn update_current_epoch(&self, max_current_epoch: HummockEpoch) -> HummockSnapshot {
+        // We only update `max_current_epoch`!
+        let prev_snapshot = self.latest_snapshot.rcu(|snapshot| HummockSnapshot {
+            committed_epoch: snapshot.committed_epoch,
+            current_epoch: max_current_epoch,
+        });
+        assert!(prev_snapshot.current_epoch < max_current_epoch);
+
+        tracing::trace!("new current epoch {}", max_current_epoch);
+        HummockSnapshot {
+            committed_epoch: prev_snapshot.committed_epoch,
+            current_epoch: max_current_epoch,
+        }
+    }
+
+    pub async fn list_change_log_epochs(
+        &self,
+        table_id: u32,
+        min_epoch: u64,
+        max_count: u32,
+    ) -> Vec<u64> {
+        let versioning = self.versioning.read().await;
+        if let Some(table_change_log) = versioning
+            .current_version
+            .table_change_log
+            .get(&TableId::new(table_id))
+        {
+            let table_change_log = table_change_log.clone();
+            table_change_log.get_epochs(min_epoch, max_count as usize)
+        } else {
+            vec![]
+        }
     }
 }
